everyone - added settings activity and improved dark mode aspect
