+++ app/src/main/java/com/example/attemptdmd/MainActivity.kt
@@ -6,27 +6,17 @@ import android.widget.*
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import kotlin.concurrent.thread
-import android.content.SharedPreferences
-import android.Manifest
-import android.content.pm.PackageManager
-import android.os.Build
-import androidx.core.app.ActivityCompat
-import androidx.core.content.ContextCompat
-import android.app.AlarmManager
-import android.app.PendingIntent
 import android.content.Context
 import android.content.Intent
 import android.util.Log
-import java.util.Calendar
 import androidx.lifecycle.lifecycleScope
 import kotlinx.coroutines.launch
 import org.json.JSONObject
 import java.net.HttpURLConnection
 import java.net.URL
-import androidx.appcompat.widget.SwitchCompat
-
-private val REQUEST_NOTIFICATION_PERMISSION = 101
-private lateinit var db: ChatDatabase
+import android.view.Menu
+import android.view.MenuItem
+import com.google.android.material.appbar.MaterialToolbar
 
 class MainActivity : AppCompatActivity() {
 
@@ -35,83 +25,25 @@ class MainActivity : AppCompatActivity() {
     private lateinit var etUserMessage: EditText
     private val chatAdapter = ChatAdapter(mutableListOf())
 
-    private lateinit var sharedPreferences: SharedPreferences
-    private lateinit var themeSwitch: SwitchCompat
-    private lateinit var notifSwitch: SwitchCompat
-
-    private var loadingItem: ChatItem.Loading? = null
+    private var lastIsDarkMode = false
 
     override fun onCreate(savedInstanceState: Bundle?) {
-        // 1) Load prefs before super.onCreate & setContentView
-        sharedPreferences = getSharedPreferences("chat_prefs", MODE_PRIVATE)
-        val isDarkMode = sharedPreferences.getBoolean("dark_mode", false)
+        val sharedPrefs = getSharedPreferences("chat_prefs", MODE_PRIVATE)
+        val isDarkMode = sharedPrefs.getBoolean("dark_mode", false)
+        lastIsDarkMode = isDarkMode
 
-        // 2) Apply the stored theme here to avoid flicker on launch
-        applyTheme(isDarkMode, immediate = false)
+        // 2) Apply dark or light theme
+        if (isDarkMode) {
+            setTheme(R.style.Theme_AttemptDMD_Dark)
+        } else {
+            setTheme(R.style.Theme_AttemptDMD)
+        }
 
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
 
-        val btnPlayAudio = findViewById<Button>(R.id.btnPlayAudio)
-        val btnStopAudio = findViewById<Button>(R.id.btnStopAudio)
-
-        btnPlayAudio.setOnClickListener {
-            startCalmingAudioService()
-        }
-
-        btnStopAudio.setOnClickListener {
-            stopCalmingAudioService()
-        }
-
-        // Initialize Room database
-        db = ChatDatabase.getDatabase(this)
-
-        // Initialize switches
-        themeSwitch = findViewById(R.id.switchTheme)
-        notifSwitch = findViewById(R.id.switchNotifications)
-
-        // Set switch states based on saved preferences
-        themeSwitch.isChecked = isDarkMode
-        val isNotificationsEnabled = sharedPreferences.getBoolean("notifications_enabled", true)
-        notifSwitch.isChecked = isNotificationsEnabled
-
-        // Listen for theme switch toggles
-        themeSwitch.setOnCheckedChangeListener { buttonView, checked ->
-            if (buttonView.isPressed) {
-                sharedPreferences.edit().putBoolean("dark_mode", checked).apply()
-                applyTheme(checked, immediate = false)
-                recreate()
-            }
-        }
-
-        // Listen for notifications switch toggles
-        notifSwitch.setOnCheckedChangeListener { _, checked ->
-            sharedPreferences.edit().putBoolean("notifications_enabled", checked).apply()
-            if (checked) {
-                scheduleDailyNotification()
-            } else {
-                cancelNotification()
-            }
-        }
-
-        // Schedule initial notification if enabled
-        if (isNotificationsEnabled) {
-            scheduleDailyNotification()
-        }
-
-        // Request notification permissions (Android 13+)
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
-            if (ContextCompat.checkSelfPermission(this, Manifest.permission.POST_NOTIFICATIONS)
-                != PackageManager.PERMISSION_GRANTED
-            ) {
-                ActivityCompat.requestPermissions(
-                    this,
-                    arrayOf(Manifest.permission.POST_NOTIFICATIONS),
-                    REQUEST_NOTIFICATION_PERMISSION
-                )
-            }
-        }
-
+        val toolbar = findViewById<MaterialToolbar>(R.id.chatToolbar)
+        setSupportActionBar(toolbar)
         // Initialize UI elements
         rvChatMessages = findViewById(R.id.rvChatMessages)
         btnSendRequest = findViewById(R.id.btnSendRequest)
@@ -131,8 +63,8 @@ class MainActivity : AppCompatActivity() {
                 chatAdapter.addItem(chatMessage)
                 etUserMessage.setText("")
 
-                loadingItem = ChatItem.Loading
-                chatAdapter.addItem(loadingItem!!)
+                val loadingItem = ChatItem.Loading
+                chatAdapter.addItem(loadingItem)
                 rvChatMessages.scrollToPosition(chatAdapter.itemCount - 1)
 
                 saveMessageToHistory(chatMessage)
@@ -143,34 +75,46 @@ class MainActivity : AppCompatActivity() {
         }
     }
 
-    /**
-     * Apply light/dark theme. If 'immediate' is false, we just call setTheme.
-     */
-    private fun applyTheme(darkMode: Boolean, immediate: Boolean) {
-        if (darkMode) {
-            setTheme(R.style.Theme_AttemptDMD_Dark)
-        } else {
-            setTheme(R.style.Theme_AttemptDMD_Light)
-        }
-        if (immediate) {
+    override fun onResume() {
+        super.onResume()
+        val sharedPrefs = getSharedPreferences("chat_prefs", MODE_PRIVATE)
+
+        val isDarkMode = sharedPrefs.getBoolean("dark_mode", false)
+        if (isDarkMode != lastIsDarkMode) {
+            lastIsDarkMode = isDarkMode
             recreate()
         }
     }
 
-    // Cancel daily notification if disabled
-    fun cancelNotification() {
-        val alarmManager = getSystemService(Context.ALARM_SERVICE) as AlarmManager
-        val intent = Intent(this, DailyNotificationReceiver::class.java)
-        val pendingIntent = PendingIntent.getBroadcast(
-            this,
-            0,
-            intent,
-            PendingIntent.FLAG_IMMUTABLE
-        )
-        alarmManager.cancel(pendingIntent)
+//    private fun reapplyTheme() {
+//        val sharedPrefs = getSharedPreferences("chat_prefs", MODE_PRIVATE)
+//        val isDarkMode = sharedPrefs.getBoolean("dark_mode", false)
+//
+//        if (isDarkMode) {
+//            setTheme(R.style.Theme_AttemptDMD_Dark)
+//        } else {
+//            setTheme(R.style.Theme_AttemptDMD)
+//        }
+//        recreate()
+//    }
+
+    override fun onCreateOptionsMenu(menu: Menu?): Boolean {
+        menuInflater.inflate(R.menu.main_menu, menu)
+        return true
     }
 
-    // Save a chat message to Room database
+    override fun onOptionsItemSelected(item: MenuItem): Boolean {
+        return when (item.itemId) {
+            R.id.action_settings -> {
+                // Launch SettingsActivity
+                val intent = Intent(this, SettingsActivity::class.java)
+                startActivity(intent)
+                true
+            }
+            else -> super.onOptionsItemSelected(item)
+        }
+    }
+
     private fun saveMessageToHistory(message: ChatItem.Message) {
         val chatMessage = ChatMessage(
             text = message.text,
@@ -178,14 +122,15 @@ class MainActivity : AppCompatActivity() {
             timestamp = System.currentTimeMillis()
         )
         lifecycleScope.launch {
-            db.chatDao().insertMessage(chatMessage)
+            ChatDatabase.getDatabase(this@MainActivity).chatDao().insertMessage(chatMessage)
         }
     }
 
-    // Load chat history from Room database
     private fun loadChatHistoryFromDb() {
         lifecycleScope.launch {
-            val messages = db.chatDao().getAllMessages()
+            val messages = ChatDatabase.getDatabase(this@MainActivity)
+                .chatDao()
+                .getAllMessages()
             runOnUiThread {
                 messages.forEach { msg ->
                     val chatItem = ChatItem.Message(msg.text, msg.isUser)
@@ -195,7 +140,6 @@ class MainActivity : AppCompatActivity() {
         }
     }
 
-    // Send chat message to the Flask API
     private fun sendChatMessage(message: String) {
         thread {
             try {
@@ -218,15 +162,14 @@ class MainActivity : AppCompatActivity() {
                     val assistantResponse = jsonResponse.optString("assistant", "No response")
 
                     runOnUiThread {
-                        loadingItem?.let { chatAdapter.removeItem(it) }
+                        chatAdapter.removeItem(ChatItem.Loading)
                         val assistantMessage = ChatItem.Message(assistantResponse, false)
                         chatAdapter.addItem(assistantMessage)
                         saveMessageToHistory(assistantMessage)
-
                         rvChatMessages.scrollToPosition(chatAdapter.itemCount - 1)
 
                         // Show immediate in-app notification
-                        val notificationHelper = NotificationHelper(this)
+                        val notificationHelper = NotificationHelper(this@MainActivity)
                         notificationHelper.showNotification("New Message", assistantResponse)
                     }
                 }
@@ -234,72 +177,10 @@ class MainActivity : AppCompatActivity() {
             } catch (e: Exception) {
                 e.printStackTrace()
                 runOnUiThread {
-                    loadingItem?.let { chatAdapter.removeItem(it) }
-                    Toast.makeText(this, "Error: ${e.message}", Toast.LENGTH_LONG).show()
+                    chatAdapter.removeItem(ChatItem.Loading)
+                    Toast.makeText(this@MainActivity, "Error: ${e.message}", Toast.LENGTH_LONG).show()
                 }
             }
         }
     }
-
-    /**
-     * Schedule a one-time exact alarm for (e.g.) 23:27.
-     * Then "DailyNotificationReceiver" will handle scheduling tomorrow's alarm.
-     */
-    private fun ensureExactAlarmPermission(grantedCallback: () -> Unit) {
-        val alarmManager = getSystemService(AlarmManager::class.java)
-
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.S) {
-            // Android 12+ requires checking canScheduleExactAlarms
-            if (!alarmManager.canScheduleExactAlarms()) {
-                // Permission not granted; show a dialog or toast, then request it:
-                Toast.makeText(this, "Need exact alarm permission", Toast.LENGTH_LONG).show()
-                val intent = Intent(android.provider.Settings.ACTION_REQUEST_SCHEDULE_EXACT_ALARM)
-                startActivity(intent)
-            } else {
-                // We already have permission
-                grantedCallback()
-            }
-        } else {
-            // Before Android 12, no special permission needed
-            grantedCallback()
-        }
-    }
-    fun scheduleDailyNotification() {
-        ensureExactAlarmPermission {
-            // This callback runs only if canScheduleExactAlarms() is true.
-            val alarmManager = getSystemService(Context.ALARM_SERVICE) as AlarmManager
-            val intent = Intent(this, DailyNotificationReceiver::class.java)
-            val pendingIntent = PendingIntent.getBroadcast(this, 0, intent, PendingIntent.FLAG_IMMUTABLE)
-
-            val calendar = Calendar.getInstance().apply {
-                set(Calendar.HOUR_OF_DAY, 18)
-                set(Calendar.MINUTE, 35)
-                set(Calendar.SECOND, 0)
-                if (timeInMillis < System.currentTimeMillis()) {
-                    add(Calendar.DAY_OF_YEAR, 1)
-                }
-            }
-
-            // This exact alarm call is now safe
-            alarmManager.setExactAndAllowWhileIdle(
-                AlarmManager.RTC_WAKEUP,
-                calendar.timeInMillis,
-                pendingIntent
-            )
-            Log.d("DailyNotificationReceiver", "Scheduled next alarm at ${calendar.time}")
-
-//            Toast.makeText(this, "Exact daily alarm scheduled at 23:36", Toast.LENGTH_SHORT).show()
-        }
-    }
-    private fun startCalmingAudioService() {
-        val intent = Intent(this, CalmingAudioService::class.java)
-        startService(intent)
-    }
-
-    // Stop
-    private fun stopCalmingAudioService() {
-        val intent = Intent(this, CalmingAudioService::class.java)
-        stopService(intent)
-    }
-
 }
