+++ app/src/main/res/drawable/bg_bubble_assistant.xml
@@ -1,8 +1,8 @@
-<?xml version="1.0" encoding="utf-8"?>
+<!-- res/drawable/bg_bubble_assistant.xml -->
 <shape xmlns:android="http://schemas.android.com/apk/res/android"
     android:shape="rectangle">
-    <solid android:color="@android:color/white"/>
-    <corners android:radius="16dp"/>
-    <stroke android:width="1dp" android:color="#DDDDDD" />
-    <padding android:left="12dp" android:top="8dp" android:right="12dp" android:bottom="8dp"/>
+
+    <solid android:color="?attr/colorBubbleAssistant" />
+
+    <corners android:radius="8dp" />
 </shape>
