+++ app/src/main/res/drawable/bg_bubble_user.xml
@@ -1,8 +1,8 @@
-<?xml version="1.0" encoding="utf-8"?>
+<!-- res/drawable/bg_bubble_user.xml -->
 <shape xmlns:android="http://schemas.android.com/apk/res/android"
     android:shape="rectangle">
-    <solid android:color="#E0F7FA"/>
-    <corners android:radius="16dp"/>
-    <stroke android:width="1dp" android:color="#B2EBF2" />
-    <padding android:left="12dp" android:top="8dp" android:right="12dp" android:bottom="8dp"/>
+
+    <solid android:color="?attr/colorBubbleUser" />
+
+    <corners android:radius="8dp" />
 </shape>
