+++ app/src/main/res/layout/activity_main.xml
@@ -6,72 +6,35 @@
     android:layout_height="match_parent"
     android:fitsSystemWindows="true">
 
-    <!-- AppBarLayout containing the Toolbar -->
     <com.google.android.material.appbar.AppBarLayout
         android:id="@+id/appbar"
         android:layout_width="match_parent"
-        android:layout_height="?attr/actionBarSize"
+        android:layout_height="wrap_content"
         android:theme="@style/ThemeOverlay.MaterialComponents.ActionBar"
         android:fitsSystemWindows="true">
 
         <com.google.android.material.appbar.MaterialToolbar
             android:id="@+id/chatToolbar"
             android:layout_width="match_parent"
-            android:layout_height="match_parent"
+            android:layout_height="?attr/actionBarSize"
             android:background="?attr/colorPrimary"
-            app:title="AttemptDMD"
-            app:subtitle="Feel free to share your thoughts"
-            app:titleTextColor="@android:color/white"
-            app:subtitleTextColor="@android:color/white" />
+            app:title="ZenChat"
+            app:titleTextColor="?attr/colorOnPrimary" />
+
+        <!-- Spacer below Toolbar -->
+        <View
+            android:layout_width="match_parent"
+            android:layout_height="5dp"
+            android:background="@android:color/transparent" />
     </com.google.android.material.appbar.AppBarLayout>
 
-    <!-- Main vertical layout that scrolls under the AppBar -->
     <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:orientation="vertical"
-        android:padding="0dp"
         app:layout_behavior="@string/appbar_scrolling_view_behavior">
 
-        <!-- A container for the two switches -->
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:orientation="vertical"
-            android:padding="16dp">
-
-            <!-- Theme (Dark Mode) Switch -->
-            <androidx.appcompat.widget.SwitchCompat
-                android:id="@+id/switchTheme"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="Dark Mode" />
-
-            <!-- Notifications Switch -->
-            <androidx.appcompat.widget.SwitchCompat
-                android:id="@+id/switchNotifications"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="Enable Daily Check-up"
-                android:layout_marginTop="8dp" />
-
-            <Button
-                android:id="@+id/btnPlayAudio"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="Play Calming Audio"
-                android:layout_marginTop="16dp" />
-
-            <Button
-                android:id="@+id/btnStopAudio"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="Stop Calming Audio"
-                android:layout_marginTop="8dp" />
-
-        </LinearLayout>
-
-        <!-- Your chat messages list -->
+        <!-- Chat messages list -->
         <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/rvChatMessages"
             android:layout_width="match_parent"
@@ -82,33 +45,32 @@
             android:paddingEnd="8dp"
             android:paddingTop="28dp"
             android:paddingBottom="64dp" />
-
     </LinearLayout>
 
-    <!-- Bottom input bar anchored to the bottom -->
+    <!-- Bottom input bar -->
     <LinearLayout
         android:id="@+id/bottom_input_bar"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:padding="8dp"
         android:gravity="center_vertical"
-        android:background="@color/white"
         android:orientation="horizontal"
-        android:layout_gravity="bottom">
+        android:layout_gravity="bottom"
+        android:background="?attr/colorSurface">
 
         <EditText
             android:id="@+id/etUserMessage"
             android:layout_width="0dp"
             android:layout_height="wrap_content"
             android:layout_weight="1"
-            android:background="@drawable/rounded_edittext"
             android:hint="Type your message..."
             android:maxLines="4"
             android:minHeight="48dp"
             android:padding="12dp"
             android:textSize="16sp"
             android:imeOptions="actionSend"
-            android:inputType="textMultiLine" />
+            android:inputType="textMultiLine"
+            android:textColor="?attr/colorOnSurface" />
 
         <ImageButton
             android:id="@+id/btnSendRequest"
@@ -117,7 +79,7 @@
             android:contentDescription="Send Message"
             android:src="@drawable/ic_send"
             android:background="?attr/selectableItemBackgroundBorderless"
-            app:tint="@color/colorPrimary" />
+            app:tint="?attr/colorPrimary" />
     </LinearLayout>
 
 </androidx.coordinatorlayout.widget.CoordinatorLayout>
