+++ app/src/main/res/layout/item_loading.xml
@@ -12,7 +12,7 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:text="Assistant is typing..."
-        android:textColor="@color/darkGray"
+        android:textColor="?attr/colorOnSurface"
         android:textSize="14sp" />
 
     <ProgressBar
