+++ app/src/main/res/values/colors.xml
@@ -1,29 +1,30 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <color name="purple_200">#FFBB86FC</color>
-    <color name="purple_500">#FF6200EE</color>
-    <color name="purple_700">#FF3700B3</color>
-    <color name="teal_200">#FF03DAC5</color>
-    <color name="teal_700">#FF018786</color>
-    <color name="colorPrimary">#6200EE</color>
-    <color name="colorPrimaryVariant">#3700B3</color>
-    <color name="colorOnPrimary">#FFFFFF</color>
-    <color name="onPrimary">#FFFFFF</color>
 
-    <color name="primary">#6200EE</color>
-    <color name="primaryVariant">#3700B3</color>
-    <color name="background">#FFFFFF</color>
-    <!-- Secondary brand color -->
-    <color name="colorSecondary">#03DAC5</color>
-    <color name="colorSecondaryVariant">#018786</color>
-    <color name="colorOnSecondary">#000000</color>
+    <!-- Light Theme Colors -->
+    <color name="colorPrimaryLight">#6200EE</color>  <!-- Deep purple -->
+    <color name="colorSecondaryLight">#03DAC6</color> <!-- Teal accent -->
+    <color name="colorOnPrimaryLight">#FFFFFF</color>
+    <color name="colorBackgroundLight">#FFFFFF</color>
+    <color name="colorOnBackgroundLight">#000000</color>
+    <color name="bubbleAssistantLight">#E0E0E0</color> <!-- A light gray -->
+    <color name="bubbleUserLight">#D1E8FF</color>      <!-- A pale bluish or any color you like -->
 
-    <!-- Other colors -->
+    <!-- Dark Theme Colors -->
+    <color name="colorPrimaryDark">#BB86FC</color>   <!-- “Desaturated” purple for dark -->
+    <color name="colorSecondaryDark">#DAB2FF</color> <!-- Teal accent still works on dark -->
+    <color name="colorOnPrimaryDark">#FFFFFF</color>
+    <color name="colorBackgroundDark">#1F1A24</color> <!--dark purple background -->
+    <color name="colorOnBackgroundDark">#FFFFFF</color>
+    <!-- Dark bubble backgrounds -->
+    <color name="bubbleAssistantDark">#423D47</color>  <!-- Dark gray -->
+    <color name="bubbleUserDark">#444466</color>
+
+    <!-- Basic neutrals you might still need -->
     <color name="white">#FFFFFF</color>
-    <color name="red">#FF0000</color>
-
     <color name="black">#000000</color>
-    <color name="lightGray">#DDDDDD</color>
-    <color name="darkGray">#888888</color>
+    <color name="lightGray">#e54041</color>
+    <color name="darkGray">#e54041</color>
 
+        <!-- Another dark color for user bubble -->
 </resources>
